CxxFlags {StanHeaders}	R Documentation

Compilation flags for StanHeaders

Description

Output the compiler or linker flags required to build with the StanHeaders package

Usage

  CxxFlags(as_character = FALSE)
  LdFlags(as_character = FALSE)

Arguments

as_character
A logical scalar that defaults to FALSE that indicates whether to return the compiler or linker flags as a character vector of length one. Otherwise, the compiler or linker flags are merely output to the screen, which is appropriate when called from a Makevars or Makevars.win file

Details

These functions are currently not exported and are typically called from a Makevars or a Makevars.win file of another package.

Value

If as_character is TRUE, then these functions return a character vector of length one. Otherwise, (which is the default) these functions return NULL invisibly after outputing the compiler or linker flags to the screen.

[Package StanHeaders version 2.32.10 Index]