StanHeaders-package {StanHeaders}	R Documentation

The Stan C++ Library Headers

Description

The Stan project implements Markov Chain Monte Carlo and optimization for statistical models. This R package provides the Stan header files so that they can easily be used by other R packages.

Details

Package: 	 StanHeaders

Type: 	 Package

Version: 	 2.9.0

Date: 	 2016-01-03

License: 	 BSD 3 clause

There are no R functions, only C++ headers. To use the Stan Math Library, it is sufficient to include LinkingTo: StanHeaders in the DESCRIPTION file of another package. The Stan Math Library is sufficient for many purposes. If, in addition, another package needs to utilize the MCMC, optimization, or parsing facilities of StanHeaders, then it is also necessary to include the src directory of StanHeaders in the other package's PKG_CPPFLAGS. For an example of such a Makevars, see https://raw.githubusercontent.com/stan-dev/rstan/develop/rstan/rstan/src/Makevars.

Author(s)

Ben Goodrich and Jiqiang Guo for the R package and the Stan Development Team for Stan. Maintainer: Ben Goodrich <benjamin.goodrich@columbia.edu>

References

http://mc-stan.org

Examples

  # No examples because there are no R functions.

[Package StanHeaders version 2.17.1 Index]